Added folder for types and utils for helper functions and server actions, WIP: Journal Page content integration with Payload
