+++ src/app/(app)/components/ScrollSpy.tsx
@@ -1,9 +1,9 @@
 "use client";
 import React, { useContext, useEffect, useState } from "react";
 import Text from "./Text";
+import { cn } from "@/lib/utils";
 import { motion } from "framer-motion";
 import * as Accordion from "@radix-ui/react-accordion";
-import { cn } from "@/lib/utils";
 import InViewContext from "../contexts/InViewContext";
 
 export type ScrollSpyType = {
