--- src/app/(app)/journal/@content/[slug]/404.tsx
@@ -1,13 +1,0 @@
-"use client";
-import { motion } from "framer-motion";
-import React from "react";
-
-const JournalPage = () => {
-  return (
-    <motion.div layout className="text top-0px">
-      Entry does not exist.
-    </motion.div>
-  );
-};
-
-export default JournalPage;
