--- src/app/(app)/journal/[slug]/page.tsx
@@ -1,5 +1,0 @@
-const Page = () => {
-  return null;
-};
-
-export default Page;
